backend/requirements.txt
fastapi>=0.115.9
uvicorn>=0.15.0
sqlmodel>=0.0.8
aiosqlite>=0.17.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
sqlalchemy>=2.0.0
